fix(cf): fix casing of Cloud Foundry
