+++ packages/cloudfoundry/src/cf.module.ts
@@ -13,23 +13,23 @@ import './pipeline/config/validation/cfTargetImpedance.validator';
 import './pipeline/config/validation/instanceSize.validator';
 import './pipeline/config/validation/requiredRoutes.validator';
 import './pipeline/stages/bakeCloudFoundryManifest/bakeCloudFoundryManifestStage';
-import './pipeline/stages/cloneServerGroup/cloudfoundryCloneServerGroupStage.module';
+import './pipeline/stages/cloneServerGroup/cloudFoundryCloneServerGroupStage.module';
 import './pipeline/stages/createServiceBindings/cloudFoundryCreateServiceBindingsStage';
-import './pipeline/stages/createServiceKey/cloudfoundryCreateServiceKeyStage.module';
+import './pipeline/stages/createServiceKey/cloudFoundryCreateServiceKeyStage.module';
 import './pipeline/stages/deleteServiceBindings/cloudFoundryDeleteServiceBindingsStage';
-import './pipeline/stages/deleteServiceKey/cloudfoundryDeleteServiceKeyStage.module';
-import './pipeline/stages/deployService/cloudfoundryDeployServiceStage.module';
-import './pipeline/stages/destroyAsg/cloudfoundryDestroyAsgStage.module';
-import './pipeline/stages/destroyService/cloudfoundryDestroyServiceStage.module';
-import './pipeline/stages/disableAsg/cloudfoundryDisableAsgStage.module';
-import './pipeline/stages/enableAsg/cloudfoundryEnableAsgStage.module';
-import './pipeline/stages/mapLoadBalancers/cloudfoundryMapLoadBalancersStage.module';
-import './pipeline/stages/resizeAsg/cloudfoundryResizeAsgStage.module';
-import './pipeline/stages/rollbackCluster/cloudfoundryRollbackClusterStage.module';
-import './pipeline/stages/runJob/cloudfoundryRunJob.module';
-import './pipeline/stages/shareService/cloudfoundryShareServiceStage.module';
-import './pipeline/stages/unmapLoadBalancers/cloudfoundryUnmapLoadBalancersStage.module';
-import './pipeline/stages/unshareService/cloudfoundryUnshareServiceStage.module';
+import './pipeline/stages/deleteServiceKey/cloudFoundryDeleteServiceKeyStage.module';
+import './pipeline/stages/deployService/cloudFoundryDeployServiceStage.module';
+import './pipeline/stages/destroyAsg/cloudFoundryDestroyAsgStage.module';
+import './pipeline/stages/destroyService/cloudFoundryDestroyServiceStage.module';
+import './pipeline/stages/disableAsg/cloudFoundryDisableAsgStage.module';
+import './pipeline/stages/enableAsg/cloudFoundryEnableAsgStage.module';
+import './pipeline/stages/mapLoadBalancers/cloudFoundryMapLoadBalancersStage.module';
+import './pipeline/stages/resizeAsg/cloudFoundryResizeAsgStage.module';
+import './pipeline/stages/rollbackCluster/cloudFoundryRollbackClusterStage.module';
+import './pipeline/stages/runJob/cloudFoundryRunJob.module';
+import './pipeline/stages/shareService/cloudFoundryShareServiceStage.module';
+import './pipeline/stages/unmapLoadBalancers/cloudFoundryUnmapLoadBalancersStage.module';
+import './pipeline/stages/unshareService/cloudFoundryUnshareServiceStage.module';
 import {
   ApplicationManagerSection,
   BoundServicesSection,
