+++ packages/cloudfoundry/src/pipeline/stages/cloneServerGroup/cloudFoundryCloneServerGroupStage.module.ts
@@ -1,12 +1,12 @@
 import type { IStage } from '@spinnaker/core';
 import { Registry } from '@spinnaker/core';
 
-import { CloudfoundryCloneServerGroupStageConfig } from './CloudfoundryCloneServerGroupStageConfig';
+import { CloudFoundryCloneServerGroupStageConfig } from './CloudFoundryCloneServerGroupStageConfig';
 
 Registry.pipeline.registerStage({
   accountExtractor: (stage: IStage) => [stage.context.credentials],
   cloudProvider: 'cloudfoundry',
-  component: CloudfoundryCloneServerGroupStageConfig,
+  component: CloudFoundryCloneServerGroupStageConfig,
   configAccountExtractor: (stage: IStage) => [stage.credentials],
   key: 'cloneServerGroup',
   provides: 'cloneServerGroup',
