+++ packages/cloudfoundry/src/pipeline/stages/createServiceKey/cloudFoundryCreateServiceKeyStage.module.ts
@@ -1,18 +1,17 @@
 import type { IStage } from '@spinnaker/core';
 import { ExecutionDetailsTasks, Registry } from '@spinnaker/core';
 
-import { CloudfoundryDeleteServiceKeyStageConfig } from './CloudfoundryDeleteServiceKeyStageConfig';
-import { CloudfoundryServiceKeyExecutionDetails } from '../../../presentation';
+import { CloudFoundryCreateServiceKeyStageConfig } from './CloudFoundryCreateServiceKeyStageConfig';
+import { CloudFoundryServiceKeyExecutionDetails } from '../../../presentation';
 
 Registry.pipeline.registerStage({
   accountExtractor: (stage: IStage) => [stage.context.credentials],
-  configAccountExtractor: (stage: IStage) => [stage.credentials],
   cloudProvider: 'cloudfoundry',
-  component: CloudfoundryDeleteServiceKeyStageConfig,
-  description: 'Delete a service key',
-  executionDetailsSections: [CloudfoundryServiceKeyExecutionDetails, ExecutionDetailsTasks],
-  key: 'deleteServiceKey',
-  label: 'Delete Service Key',
+  component: CloudFoundryCreateServiceKeyStageConfig,
+  description: 'Create a service key',
+  executionDetailsSections: [CloudFoundryServiceKeyExecutionDetails, ExecutionDetailsTasks],
+  key: 'createServiceKey',
+  label: 'Create Service Key',
   validators: [
     { type: 'requiredField', fieldName: 'credentials', fieldLabel: 'account' },
     { type: 'requiredField', fieldName: 'region', preventSave: true },
