+++ packages/cloudfoundry/src/pipeline/stages/deleteServiceKey/cloudFoundryDeleteServiceKeyStage.module.ts
@@ -1,17 +1,18 @@
 import type { IStage } from '@spinnaker/core';
 import { ExecutionDetailsTasks, Registry } from '@spinnaker/core';
 
-import { CloudfoundryCreateServiceKeyStageConfig } from './CloudfoundryCreateServiceKeyStageConfig';
-import { CloudfoundryServiceKeyExecutionDetails } from '../../../presentation';
+import { CloudFoundryDeleteServiceKeyStageConfig } from './CloudFoundryDeleteServiceKeyStageConfig';
+import { CloudFoundryServiceKeyExecutionDetails } from '../../../presentation';
 
 Registry.pipeline.registerStage({
   accountExtractor: (stage: IStage) => [stage.context.credentials],
+  configAccountExtractor: (stage: IStage) => [stage.credentials],
   cloudProvider: 'cloudfoundry',
-  component: CloudfoundryCreateServiceKeyStageConfig,
-  description: 'Create a service key',
-  executionDetailsSections: [CloudfoundryServiceKeyExecutionDetails, ExecutionDetailsTasks],
-  key: 'createServiceKey',
-  label: 'Create Service Key',
+  component: CloudFoundryDeleteServiceKeyStageConfig,
+  description: 'Delete a service key',
+  executionDetailsSections: [CloudFoundryServiceKeyExecutionDetails, ExecutionDetailsTasks],
+  key: 'deleteServiceKey',
+  label: 'Delete Service Key',
   validators: [
     { type: 'requiredField', fieldName: 'credentials', fieldLabel: 'account' },
     { type: 'requiredField', fieldName: 'region', preventSave: true },
