+++ packages/cloudfoundry/src/pipeline/stages/destroyAsg/cloudFoundryDestroyAsgStage.module.ts
@@ -1,15 +1,20 @@
 import type { IStage } from '@spinnaker/core';
 import { Registry } from '@spinnaker/core';
-import { CloudfoundryAsgStageConfig } from '../../../presentation';
+import { CloudFoundryAsgStageConfig } from '../../../presentation';
 
 Registry.pipeline.registerStage({
   accountExtractor: (stage: IStage) => [stage.context.credentials],
   cloudProvider: 'cloudfoundry',
-  component: CloudfoundryAsgStageConfig,
+  component: CloudFoundryAsgStageConfig,
   configAccountExtractor: (stage: IStage) => [stage.credentials],
-  key: 'disableServerGroup',
-  provides: 'disableServerGroup',
+  key: 'destroyServerGroup',
+  provides: 'destroyServerGroup',
   validators: [
+    {
+      type: 'cfTargetImpedance',
+      message:
+        'This pipeline will attempt to destroy a server group without deploying a new version into the same cluster.',
+    },
     { type: 'requiredField', fieldName: 'cluster' },
     { type: 'requiredField', fieldName: 'target' },
     { type: 'requiredField', fieldName: 'regions' },
