+++ packages/cloudfoundry/src/pipeline/stages/unmapLoadBalancers/cloudFoundryUnmapLoadBalancersStage.module.ts
@@ -1,16 +1,16 @@
 import type { IStage } from '@spinnaker/core';
 import { ExecutionDetailsTasks, Registry } from '@spinnaker/core';
-import { CloudfoundryLoadBalancersExecutionDetails, CloudfoundryLoadBalancersStageConfig } from '../../../presentation';
+import { CloudFoundryLoadBalancersExecutionDetails, CloudFoundryLoadBalancersStageConfig } from '../../../presentation';
 
 Registry.pipeline.registerStage({
   accountExtractor: (stage: IStage) => [stage.context.credentials],
   configAccountExtractor: (stage: IStage) => [stage.credentials],
   cloudProvider: 'cloudfoundry',
-  component: CloudfoundryLoadBalancersStageConfig,
-  description: 'Map a load balancer',
-  executionDetailsSections: [CloudfoundryLoadBalancersExecutionDetails, ExecutionDetailsTasks],
-  key: 'mapLoadBalancers',
-  label: 'Map Load Balancer',
+  component: CloudFoundryLoadBalancersStageConfig,
+  description: 'Unmap a load balancer',
+  executionDetailsSections: [CloudFoundryLoadBalancersExecutionDetails, ExecutionDetailsTasks],
+  key: 'unmapLoadBalancers',
+  label: 'Unmap Load Balancer',
   validators: [
     { type: 'requiredField', preventSave: true, fieldName: 'cluster' },
     { type: 'requiredField', preventSave: true, fieldName: 'credentials', fieldLabel: 'account' },
