+++ packages/cloudfoundry/src/pipeline/stages/unshareService/CloudFoundryUnshareServiceExecutionDetails.tsx
@@ -4,13 +4,12 @@ import React from 'react';
 import type { IExecutionDetailsSectionProps } from '@spinnaker/core';
 import { AccountTag, ExecutionDetailsSection, StageExecutionLogs, StageFailureMessage } from '@spinnaker/core';
 
-export function CloudfoundryShareServiceExecutionDetails(props: IExecutionDetailsSectionProps) {
+export function CloudFoundryUnshareServiceExecutionDetails(props: IExecutionDetailsSectionProps) {
   const { stage } = props;
   const { context } = stage;
   const account = get(context, 'service.account', undefined);
-  const region = get(context, 'service.region', undefined);
   const serviceInstanceName = get(context, 'serviceInstanceName', undefined);
-  const shareToRegions = get(context, 'shareToRegions', undefined);
+  const unshareFromRegions = get(context, 'unshareFromRegions', undefined);
   return (
     <ExecutionDetailsSection name={props.name} current={props.current}>
       <div className="step-section-details">
@@ -21,19 +20,14 @@ export function CloudfoundryShareServiceExecutionDetails(props: IExecutionDetail
               <dd>
                 <AccountTag account={account} />
               </dd>
-              <dt>Region</dt>
-              <dd>
-                {region}
-                <br />
-              </dd>
               <dt>Service Instance Name</dt>
               <dd>
                 {serviceInstanceName}
                 <br />
               </dd>
-              <dt>Sharing Regions</dt>
+              <dt>Unsharing Regions</dt>
               <dd>
-                {shareToRegions.join(', ')}
+                {unshareFromRegions.join(', ')}
                 <br />
               </dd>
             </dl>
@@ -47,6 +41,6 @@ export function CloudfoundryShareServiceExecutionDetails(props: IExecutionDetail
 }
 
 // eslint-disable-next-line
-export namespace CloudfoundryShareServiceExecutionDetails {
-  export const title = 'cloudfoundryShareServiceConfig';
+export namespace CloudFoundryUnshareServiceExecutionDetails {
+  export const title = 'cloudfoundryUnshareServiceConfig';
 }
