+++ packages/cloudfoundry/src/pipeline/stages/unshareService/CloudFoundryUnshareServiceStageConfig.tsx
@@ -7,15 +7,14 @@ import { takeUntil } from 'rxjs/operators';
 import type { IAccount, IRegion, IStageConfigProps } from '@spinnaker/core';
 import { AccountService, ReactSelectInput, StageConfigField, TextInput } from '@spinnaker/core';
 
-interface ICloudfoundryShareServiceStageConfigState {
+interface ICloudFoundryShareServiceStageConfigState {
   accounts: IAccount[];
   regions: string[];
-  shareToRegionsList: string[];
 }
 
-export class CloudfoundryShareServiceStageConfig extends React.Component<
+export class CloudFoundryUnshareServiceStageConfig extends React.Component<
   IStageConfigProps,
-  ICloudfoundryShareServiceStageConfigState
+  ICloudFoundryShareServiceStageConfigState
 > {
   private destroy$ = new Subject();
 
@@ -25,7 +24,6 @@ export class CloudfoundryShareServiceStageConfig extends React.Component<
     this.state = {
       accounts: [],
       regions: [],
-      shareToRegionsList: [],
     };
   }
 
@@ -47,13 +45,9 @@ export class CloudfoundryShareServiceStageConfig extends React.Component<
     observableFrom(AccountService.getRegionsForAccount(this.props.stage.credentials))
       .pipe(takeUntil(this.destroy$))
       .subscribe((regionList: IRegion[]) => {
-        const { region } = this.props.stage;
         const regions = regionList.map((r) => r.name);
         regions.sort((a, b) => a.localeCompare(b));
         this.setState({ regions });
-        if (region) {
-          this.clearAndResetShareToRegionList(region, regions);
-        }
       });
   };
 
@@ -61,43 +55,25 @@ export class CloudfoundryShareServiceStageConfig extends React.Component<
     this.props.updateStageField({ serviceInstanceName: event.target.value });
   };
 
-  private clearAndResetShareToRegionList = (region: string, regions: string[]) => {
-    this.setState({ shareToRegionsList: regions.filter((r) => r !== region) });
-  };
-
   private accountUpdated = (option: Option<string>) => {
     const credentials = option.target.value;
-    this.setState({
-      regions: [],
-      shareToRegionsList: [],
-    });
+    this.setState({ regions: [] });
     this.props.updateStageField({
       credentials,
-      region: '',
-      shareToRegions: [],
+      unshareFromRegions: [],
     });
     if (credentials) {
       this.clearAndReloadRegions();
     }
   };
 
-  private regionUpdated = (option: Option<string>) => {
-    const region = option.target.value;
-    this.setState({ shareToRegionsList: [] });
-    this.props.updateStageField({
-      region,
-      shareToRegions: [],
-    });
-    this.clearAndResetShareToRegionList(region, this.state.regions);
-  };
-
-  private shareToRegionsUpdated = (option: Option<string>) => {
-    this.props.updateStageField({ shareToRegions: option.map((o: Option) => o.value) });
+  private unshareFromRegionsUpdated = (option: Option<string>) => {
+    this.props.updateStageField({ unshareFromRegions: option.map((o: Option) => o.value) });
   };
 
   public render() {
-    const { credentials, region, serviceInstanceName, shareToRegions } = this.props.stage;
-    const { accounts, regions, shareToRegionsList } = this.state;
+    const { credentials, serviceInstanceName, unshareFromRegions } = this.props.stage;
+    const { accounts, regions } = this.state;
 
     return (
       <div className="form-horizontal">
@@ -109,9 +85,6 @@ export class CloudfoundryShareServiceStageConfig extends React.Component<
             stringOptions={accounts.map((it) => it.name)}
           />
         </StageConfigField>
-        <StageConfigField label="Region">
-          <ReactSelectInput clearable={false} onChange={this.regionUpdated} value={region} stringOptions={regions} />
-        </StageConfigField>
         <StageConfigField label="Service Instance Name">
           <TextInput
             type="text"
@@ -120,19 +93,19 @@ export class CloudfoundryShareServiceStageConfig extends React.Component<
             value={serviceInstanceName}
           />
         </StageConfigField>
-        <StageConfigField label="Share To Regions">
+        <StageConfigField label="Unshare From Regions">
           <Select
             options={
-              shareToRegionsList &&
-              shareToRegionsList.map((r: string) => ({
+              regions &&
+              regions.map((r: string) => ({
                 label: r,
                 value: r,
               }))
             }
             multi={true}
             clearable={false}
-            value={shareToRegions}
-            onChange={this.shareToRegionsUpdated}
+            value={unshareFromRegions}
+            onChange={this.unshareFromRegionsUpdated}
           />
         </StageConfigField>
       </div>
