+++ packages/cloudfoundry/src/pipeline/stages/unshareService/cloudFoundryUnshareServiceStage.module.ts
@@ -1,21 +1,21 @@
 import type { IStage } from '@spinnaker/core';
 import { ExecutionDetailsTasks, Registry } from '@spinnaker/core';
 
-import { CloudfoundryShareServiceExecutionDetails } from './CloudfoundryShareServiceExecutionDetails';
-import { CloudfoundryShareServiceStageConfig } from './CloudfoundryShareServiceStageConfig';
+import { CloudFoundryUnshareServiceExecutionDetails } from './CloudFoundryUnshareServiceExecutionDetails';
+import { CloudFoundryUnshareServiceStageConfig } from './CloudFoundryUnshareServiceStageConfig';
 
 Registry.pipeline.registerStage({
   accountExtractor: (stage: IStage) => [stage.context.credentials],
   configAccountExtractor: (stage: IStage) => [stage.credentials],
-  provides: 'shareService',
-  key: 'shareService',
+  provides: 'unshareService',
+  key: 'unshareService',
   cloudProvider: 'cloudfoundry',
-  component: CloudfoundryShareServiceStageConfig,
-  executionDetailsSections: [CloudfoundryShareServiceExecutionDetails, ExecutionDetailsTasks],
+  component: CloudFoundryUnshareServiceStageConfig,
+  controller: 'cfUnshareServiceStageCtrl',
+  executionDetailsSections: [CloudFoundryUnshareServiceExecutionDetails, ExecutionDetailsTasks],
   supportsCustomTimeout: true,
   validators: [
     { type: 'requiredField', fieldName: 'credentials', fieldLabel: 'account' },
-    { type: 'requiredField', fieldName: 'region' },
     { type: 'requiredField', fieldName: 'serviceInstanceName', preventSave: true },
   ],
 });
