+++ packages/cloudfoundry/src/presentation/pipeline/stages/CloudFoundryLoadBalancersExecutionDetails.tsx
@@ -4,13 +4,13 @@ import React from 'react';
 import type { IExecutionDetailsSectionProps } from '@spinnaker/core';
 import { AccountTag, ExecutionDetailsSection, StageExecutionLogs, StageFailureMessage } from '@spinnaker/core';
 
-export function CloudfoundryServiceKeyExecutionDetails(props: IExecutionDetailsSectionProps) {
+export function CloudFoundryLoadBalancersExecutionDetails(props: IExecutionDetailsSectionProps) {
   const { stage } = props;
   const { context } = stage;
-  const account = get(context, 'service.account', undefined);
-  const region = get(context, 'service.region', undefined);
-  const serviceInstanceName = get(context, 'serviceInstanceName', undefined);
-  const serviceKeyName = get(context, 'serviceKeyName', undefined);
+  const account = get(context, 'credentials', undefined);
+  const region = get(context, 'region', undefined);
+  const loadBalancerNames = get(context, 'loadBalancerNames', undefined);
+  const serverGroupName = get(context, 'serverGroupName', undefined);
   return (
     <ExecutionDetailsSection name={props.name} current={props.current}>
       <div className="step-section-details">
@@ -26,14 +26,14 @@ export function CloudfoundryServiceKeyExecutionDetails(props: IExecutionDetailsS
                 {region}
                 <br />
               </dd>
-              <dt>Service Instance Name</dt>
+              <dt>Server Group Name</dt>
               <dd>
-                {serviceInstanceName}
+                {serverGroupName}
                 <br />
               </dd>
-              <dt>Service Key Name</dt>
+              <dt>Load Balancers</dt>
               <dd>
-                {serviceKeyName}
+                {loadBalancerNames.join(', ')}
                 <br />
               </dd>
             </dl>
@@ -46,7 +46,8 @@ export function CloudfoundryServiceKeyExecutionDetails(props: IExecutionDetailsS
   );
 }
 
+// TODO: refactor this to not use namespace
 // eslint-disable-next-line
-export namespace CloudfoundryServiceKeyExecutionDetails {
-  export const title = 'cloudfoundryServiceKeyConfig';
+export namespace CloudFoundryLoadBalancersExecutionDetails {
+  export const title = 'cloudfoundryLoadBalancersConfig';
 }
