+++ packages/cloudfoundry/src/presentation/pipeline/stages/CloudFoundryServiceKeyExecutionDetails.tsx
@@ -4,13 +4,13 @@ import React from 'react';
 import type { IExecutionDetailsSectionProps } from '@spinnaker/core';
 import { AccountTag, ExecutionDetailsSection, StageExecutionLogs, StageFailureMessage } from '@spinnaker/core';
 
-export function CloudfoundryLoadBalancersExecutionDetails(props: IExecutionDetailsSectionProps) {
+export function CloudFoundryServiceKeyExecutionDetails(props: IExecutionDetailsSectionProps) {
   const { stage } = props;
   const { context } = stage;
-  const account = get(context, 'credentials', undefined);
-  const region = get(context, 'region', undefined);
-  const loadBalancerNames = get(context, 'loadBalancerNames', undefined);
-  const serverGroupName = get(context, 'serverGroupName', undefined);
+  const account = get(context, 'service.account', undefined);
+  const region = get(context, 'service.region', undefined);
+  const serviceInstanceName = get(context, 'serviceInstanceName', undefined);
+  const serviceKeyName = get(context, 'serviceKeyName', undefined);
   return (
     <ExecutionDetailsSection name={props.name} current={props.current}>
       <div className="step-section-details">
@@ -26,14 +26,14 @@ export function CloudfoundryLoadBalancersExecutionDetails(props: IExecutionDetai
                 {region}
                 <br />
               </dd>
-              <dt>Server Group Name</dt>
+              <dt>Service Instance Name</dt>
               <dd>
-                {serverGroupName}
+                {serviceInstanceName}
                 <br />
               </dd>
-              <dt>Load Balancers</dt>
+              <dt>Service Key Name</dt>
               <dd>
-                {loadBalancerNames.join(', ')}
+                {serviceKeyName}
                 <br />
               </dd>
             </dl>
@@ -46,8 +46,7 @@ export function CloudfoundryLoadBalancersExecutionDetails(props: IExecutionDetai
   );
 }
 
-// TODO: refactor this to not use namespace
 // eslint-disable-next-line
-export namespace CloudfoundryLoadBalancersExecutionDetails {
-  export const title = 'cloudfoundryLoadBalancersConfig';
+export namespace CloudFoundryServiceKeyExecutionDetails {
+  export const title = 'cloudfoundryServiceKeyConfig';
 }
