+++ packages/cloudfoundry/src/presentation/pipeline/stages/index.ts
@@ -1,5 +1,5 @@
-export * from './CloudfoundryAsgStageConfig';
-export * from './CloudfoundryLoadBalancersExecutionDetails';
-export * from './CloudfoundryLoadBalancersStageConfig';
-export * from './CloudfoundryServiceExecutionDetails';
-export * from './CloudfoundryServiceKeyExecutionDetails';
+export * from './CloudFoundryAsgStageConfig';
+export * from './CloudFoundryLoadBalancersExecutionDetails';
+export * from './CloudFoundryLoadBalancersStageConfig';
+export * from './CloudFoundryServiceExecutionDetails';
+export * from './CloudFoundryServiceKeyExecutionDetails';
